Restructure discord class and add basic bot config
