+++ deal_finder/constants.py
@@ -1,1 +1,2 @@
 ERROR_NOTIFICATION_EMAIL_ADDRESSES = []
+BOT_COMMAND_PREFIX = "!"
