+++ deal_finder/discord_client/__init__.py
@@ -1,0 +1,1 @@
+from discord_client.client import DiscordWrapper
